+++ Selfienator/src/selfienator/ICommands.java
@@ -3,8 +3,11 @@ package selfienator;
 public interface ICommands {
 
 	public void takeShot(ShotSettings settings);
-	public void moveTo(int x, int y);
-	public void moveTo(int x, int y, Directions dir);
-	public void moveTo(int x, int y, Directions dirX, Directions dirY);
-	
+
+	public void moveTo(double x, double y);
+
+	public void moveTo(double x, double y, Directions dir);
+
+	public void moveTo(double x, double y, Directions dirX, Directions dirY);
+
 }
